IMPORTANT COMMIT: starting to clean and refactore the code. The current structure of the file is now:
core: contain all the core function which allows to manipulate genotype and haplotype.
frequency_evolution: contain the algorithm itself for the evolution of the genotype in the population and related function
fitness_determination: all function needed to find the fitness of an individual
sex_determination: all function needed to find the sex of an individual
recombination: all function used to perform recombination
interface_function: function which are usefull to user. All function in this file should be documented.
interface_subfunction: function which are not usefull for the user but called in interface.
plot: all function to plot something. All function in this file should be documented
utilities_function: function which are not related at all directly to the package, but still usefull.